improvement - refactoring
