+++ helios/extraction_engine/agent_specs/sec_10k.md
@@ -12,7 +12,7 @@
 - **Report Metadata:** Exact submission date, the exact SEC form type (e.g., 10-K, 10-K/A), the company ticker symbol, and primary currency.
 {{ business_and_risk }}
 - **Financial Statements:** Extract the FULL, line-by-line quantitative tables for the following statements. Include geographic or product-segment breakdowns if explicitly reported:
-    - Income Statement (including also diluted shares oustanding if available)
+    - Income Statement (including also diluted shares outstanding if available)
     - Balance Sheet (including also net debt if available)
     - Cash Flow Statement (including also free cash flow if available)
 - **Revenue & Cost Structure:** Extract the exact categorical breakdown of revenue streams and the primary drivers of Cost of Goods Sold (COGS) / Operating Expenses. Do not summarize; use the company's exact terminology.
