+++ helios/synthesis_engine/business_overview_synthesizer.py
@@ -20,7 +20,8 @@ from pathlib import Path
 
 from helios.config import GEMINI, SYNTHESIS_AGENT_SPECS_DIR, OutputDir
 from helios.extraction_engine.edgar.domain import EdgarFormType
-from helios.utils.commons import DeepResearchAnalyser, read_dir_files
+from helios.utils.deep_research_analyser import DeepResearchAnalyser
+from helios.utils.commons import format_dossier_section, read_dir_files
 
 logger = logging.getLogger(__name__)
 
@@ -50,57 +51,43 @@ class BusinessOverviewSynthesizer(DeepResearchAnalyser):
     def _build_context(self) -> str:
         """Load upstream pipeline outputs as context for Deep Research."""
         ticker_dir = os.path.join(self.output_base_dir, self.ticker)
-        sections: list[str] = []
 
         # Latest 10-K only
         edgar_base = os.path.join(ticker_dir, OutputDir.EDGAR_SUMMARIZED)
         form_dir = os.path.join(edgar_base, EdgarFormType.ANNUAL_REPORT)
+        edgar_docs: list[tuple[str, str]] = []
         if os.path.isdir(form_dir):
             files = sorted(Path(form_dir).glob("*.json"), reverse=True)[:1]
-            edgar_docs = []
             for fp in files:
                 content = fp.read_text(encoding="utf-8")
                 if content:
                     edgar_docs.append((f"EDGAR {EdgarFormType.ANNUAL_REPORT}/{fp.name} (latest only)", content))
-            if edgar_docs:
-                self._add_section(sections, "EDGAR FILINGS (10-K latest only)", edgar_docs)
 
-        # Quantitative Baseline
-        qbc_docs = read_dir_files(os.path.join(ticker_dir, OutputDir.QUANTITATIVE_BASELINE), "*.yaml")
-        if qbc_docs:
-            self._add_section(sections, "QUANTITATIVE BASELINE PAYLOAD (YAML)", qbc_docs)
+        sections = [
+            format_dossier_section("EDGAR FILINGS (10-K latest only)", edgar_docs),
+            format_dossier_section(
+                "QUANTITATIVE BASELINE PAYLOAD (YAML)",
+                read_dir_files(os.path.join(ticker_dir, OutputDir.QUANTITATIVE_BASELINE), "*.yaml"),
+            ),
+            format_dossier_section(
+                "NARRATIVE VALIDATION PAYLOAD",
+                read_dir_files(os.path.join(ticker_dir, OutputDir.NARRATIVE_VALIDATION), "*.md"),
+            ),
+            format_dossier_section(
+                "SECTOR ANALYSIS PAYLOAD",
+                read_dir_files(os.path.join(ticker_dir, OutputDir.SECTOR_ANALYSIS), "*.md"),
+            ),
+            format_dossier_section(
+                "EARNINGS CALL SYNTHESIS",
+                read_dir_files(os.path.join(ticker_dir, OutputDir.EARNINGS_CALLS), "*.md"),
+            ),
+        ]
 
-        # Narrative Validation
-        nv_docs = read_dir_files(os.path.join(ticker_dir, OutputDir.NARRATIVE_VALIDATION), "*.md")
-        if nv_docs:
-            self._add_section(sections, "NARRATIVE VALIDATION PAYLOAD", nv_docs)
-
-        # Sector Analysis
-        se_docs = read_dir_files(os.path.join(ticker_dir, OutputDir.SECTOR_ANALYSIS), "*.md")
-        if se_docs:
-            self._add_section(sections, "SECTOR ANALYSIS PAYLOAD", se_docs)
-
-        # Earnings Calls
-        ec_docs = read_dir_files(os.path.join(ticker_dir, OutputDir.EARNINGS_CALLS), "*.md")
-        if ec_docs:
-            self._add_section(sections, "EARNINGS CALL SYNTHESIS", ec_docs)
-
-        if not sections:
+        dossier = "\n\n".join(s for s in sections if s)
+        if not dossier:
             raise FileNotFoundError(
                 f"No source documents found for {self.ticker}. Run the extraction and synthesis engines first."
             )
 
-        logger.info(f"[BusinessOverviewSynthesizer] Loaded context ({sum(len(s) for s in sections):,} chars).")
-        return "\n".join(sections)
-
-    @staticmethod
-    def _add_section(sections: list[str], heading: str, docs: list[tuple[str, str]]) -> None:
-        """Append a labeled document group to sections."""
-        separator = "=" * 60
-        prefix = "" if not sections else "\n"
-        sections.append(prefix + separator)
-        sections.append(heading)
-        sections.append(separator)
-        for label, content in docs:
-            sections.append(f"\n--- {label} ---\n")
-            sections.append(content)
+        logger.info(f"[BusinessOverviewSynthesizer] Loaded context ({len(dossier):,} chars).")
+        return dossier
