+++ helios/synthesis_engine/narrative_validator.py
@@ -16,7 +16,8 @@ import logging
 import os
 
 from helios.config import GEMINI, SYNTHESIS_AGENT_SPECS_DIR, OutputDir
-from helios.utils.commons import DossierAnalyser, current_quarter
+from helios.utils.commons import format_dossier_section
+from helios.utils.dossier_analyser import DossierAnalyser
 
 logger = logging.getLogger(__name__)
 
@@ -39,20 +40,17 @@ class NarrativeValidator(DossierAnalyser):
         return GEMINI.narrative_validator_temperature
 
     def _build_output_path(self) -> str:
-        year, quarter = current_quarter()
+        year, quarter = self._current_quarter()
         filename = f"report_{year}-Q{quarter}.md"
         return os.path.join(self._ticker_dir(), OutputDir.NARRATIVE_VALIDATION, filename)
 
     def _compile_dossier(self) -> str:
-        sections: list[str] = []
-        self._add_dossier_section(sections, "SECTOR ANALYSIS", self._collect_files(OutputDir.SECTOR_ANALYSIS, "*.md"))
-        self._add_dossier_section(
-            sections, "EARNINGS CALL SYNTHESIS", self._collect_files(OutputDir.EARNINGS_CALLS, "*.md")
-        )
-        self._add_dossier_section(sections, "EDGAR FILINGS (latest first)", self._collect_edgar_filings())
-
-        if not sections:
-            raise FileNotFoundError(
-                f"No source documents found for {self.ticker}. " f"Run the extraction engine first."
-            )
-        return "\n".join(sections)
+        sections = [
+            format_dossier_section("SECTOR ANALYSIS", self._collect_files(OutputDir.SECTOR_ANALYSIS, "*.md")),
+            format_dossier_section("EARNINGS CALL SYNTHESIS", self._collect_files(OutputDir.EARNINGS_CALLS, "*.md")),
+            format_dossier_section("EDGAR FILINGS (latest first)", self._collect_edgar_filings()),
+        ]
+        dossier = "\n\n".join(s for s in sections if s)
+        if not dossier:
+            raise FileNotFoundError(f"No source documents found for {self.ticker}. Run the extraction engine first.")
+        return dossier
